+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 and this project attempts to adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED] - tbc
+
+## Added
+
+- `geometry.absolute_bearing()`
+
+## Changed
+
+- Rename `maths` module to `geometry`
+
 ## [0.1.0] - 2025-11-13
 
 Initial release